chore(warnings): progress commit - clear C4018, C4267 categories
Promotes int loop counters to uint32_t to match unsigned bounds (6 sites) and adds explicit static_cast<T>(...) for size_t -> uint32_t/int/uint8_t narrowings (30 sites). No public API changes; build clean, all categories of these codes are at zero. Updates WARNINGS.md to reflect live state.

+++ src/model/src/database/rocprofvis_db.cpp
@@ -396,7 +396,7 @@ size_t Database::GetInfoTableNumRows(rocprofvis_dm_table_t object){
 const char* Database::GetInfoTableColumnName(rocprofvis_dm_table_t object, size_t column_index){
     TableCache* table = (TableCache*)object;
     ROCPROFVIS_ASSERT_MSG_RETURN(table, ERROR_TABLE_CANNOT_BE_NULL, 0);
-    return table->GetColumnName(column_index);
+    return table->GetColumnName(static_cast<uint32_t>(column_index));
 }
 rocprofvis_dm_table_row_t Database::GetInfoTableRowHandle(rocprofvis_dm_table_t object, size_t row_index){
     TableCache* table = (TableCache*)object;
